+++ src/cli/cfgen/bin.ts
@@ -1,5 +1,0 @@
-#!/usr/bin/env bun
-import { execute } from "../core/execute";
-import { createGenEnv } from "./cf-env-command";
-
-await execute(createGenEnv());